Add integration tests
The integration tests require a running PostGIS instance and spatial
database. This adds two new tox environments to allow for running the
unit tests without needing to have PostGIS set up. The full suite of
tests will be run during CI.
