+++ tox.ini
@@ -1,27 +1,20 @@
 [tox]
-envlist = py27,py35,setup,coverage
+envlist = py27,py35,py{27,35}-integration,coverage,flake8
 
 [testenv]
-passenv = HOME
-deps = pipenv
+passenv = HOME POSTGIS_DB
+deps =
+    pipenv
+    python-dotenv
+    py27: mock
+    {coverage,coveralls}: pytest-cov
+    coveralls: coveralls
+setenv =
+    {integration,coverage}: PYTEST_ARGS=tests/integration
+    coverage: PYTEST_COV="--cov=slingshot"
 commands =
     pipenv install --dev --system
-    py.test tests {posargs:--tb=short}
-
-[testenv:setup]
-basepython = python3.5
-deps =
-commands = python setup.py test
-
-[testenv:coverage]
-basepython = python3.5
-commands =
-    pipenv install --dev --system
-    py.test --cov=slingshot {posargs}
-deps =
-    mock
-    pytest-cov
-    {[testenv]deps}
+    py.test tests/unit {env:PYTEST_ARGS:} {env:PYTEST_COV:} {posargs:--tb=short}
 
 [testenv:flake8]
 basepython = python3.5
@@ -29,12 +22,9 @@ deps = flake8
 commands = flake8
 
 [testenv:coveralls]
-passenv = TRAVIS TRAVIS_JOB_ID TRAVIS_BRANCH
+passenv = TRAVIS TRAVIS_JOB_ID TRAVIS_BRANCH POSTGIS_DB
 basepython = python3.5
-deps =
-    coveralls
-    {[testenv:coverage]deps}
 commands =
     pipenv install --dev --system
-    py.test --cov=slingshot
+    py.test tests --cov=slingshot
     coveralls
